% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crs.R
\name{crs_input}
\alias{crs_input}
\title{Extract 'input' value}
\usage{
crs_input(x, ...)
}
\arguments{
\item{x}{object with 'input' value}

\item{...}{ignored}
}
\value{
character (or NA)
}
\description{
Obtain the 'input' string from an object, if it has one. Supported inputs
include sf (>= 0.8-1 - probably).
}
\examples{
crs_input(sfx)
x <- sfx
attr(x$geom, "crs")$epsg <- NA ## oh no we lost it
crs_epsg(x)
}
\references{
\href{http://r-spatial.github.io/sf/}{sf}
}
\seealso{
\code{\link[=crs_wkt2]{crs_wkt2()}} \code{\link[=crs_proj]{crs_proj()}} \code{\link[=crs_epsg]{crs_epsg()}}
}
