% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsmeta-package.R
\name{crs_wkt}
\alias{crs_wkt}
\title{Catch for early users}
\usage{
crs_wkt(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\value{
nothing, an error
}
\description{
Catch for early users
}
