% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crs.R
\name{crs_epsg}
\alias{crs_epsg}
\title{Extract 'EPSG' value}
\usage{
crs_epsg(x, ...)
}
\arguments{
\item{x}{object with 'EPSG' value}

\item{...}{ignored}
}
\value{
integer (or NA)
}
\description{
Obtain the 'EPSG' string from an object, if it has one. Supported inputs
include sf.
}
\seealso{
crs_wkt2 crs_proj
}
