% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crs.R
\name{crs_proj}
\alias{crs_proj}
\title{Extract PROJ4 string}
\usage{
crs_proj(x, ...)
}
\arguments{
\item{x}{object with PROJ4 string}

\item{...}{ignored}
}
\value{
character string (or NA)
}
\description{
Obtain the PROJ4 string from an object, if it has one. Supported inputs
include raster, sf, sp, and silicate.
}
\seealso{
crs_epsg crs_wkt2
}
