% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crs.R
\name{crs_wkt}
\alias{crs_wkt}
\title{Extract 'WKT2' string}
\usage{
crs_wkt(x, ...)
}
\arguments{
\item{x}{object with 'WKT2' string}

\item{...}{ignored}
}
\value{
character string (or NA)
}
\description{
Obtain the 'WKT2' string from an object, if it has one. Supported inputs
include sp (and sf soon).
}
\section{Warning}{
 For sp WKT2 only, PROJ6 and beyond
}

\seealso{
crs_epsg crs_proj
}
