% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsmeta-package.R
\docType{package}
\name{crsmeta-package}
\alias{crsmeta}
\alias{crsmeta-package}
\title{crsmeta: Extract Coordinate System Metadata}
\description{
Obtain coordinate system metadata from various formats. This is
 purely for getting simple metadata from in-memory formats, please use other
 tools for out of memory data sources.

Obtain coordinate system metadata from various in-memory R formats.
}
\details{
There are a few functions to return a character string of length one of each
main kind of projection string.
\tabular{ll}{
\code{\link{crs_epsg}}         \tab short numeric EPSG lookup \cr
\code{\link{crs_proj}}        \tab short PROJ4 style string  \cr
\code{\link{crs_wkt2}}        \tab the full parameter set as WKT2 string \cr
}
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})

}
\keyword{internal}
